Move Lipmaa decomposition to df package
This commit moves lipmaa decomposition functionality
from common package to df package, since this is the
only place where it is used. It is now unexported.
Some tests were renamed for consistency and brevity,
and an import of assert package was fixed.
